+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v51.0.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+
+* New `GridModel` method `ensureRecordsVisibleAsync` accepts one or more store records or IDs and
+  scrolls to make them visible in the grid. Existing method `ensureSelectionVisibleAsync`
+  internally calls the new method.
+
 ## v50.0.0 - 2022-07-12
 
 ### 🎁 New Features