+++ CHANGELOG.md
@@ -5,8 +5,7 @@
 ### 🎁 New Features
 
 * New `GridModel` method `ensureRecordsVisibleAsync` accepts one or more store records or IDs and
-  scrolls to make them visible in the grid. Existing method `ensureSelectionVisibleAsync`
-  internally calls the new method.
+  scrolls to make them visible in the grid.
 
 ## v50.0.0 - 2022-07-12